Amend tests and move onInputChange call
onInputChange is now called from `handleInputChange`, which is less
surprising than `componentDidUpdate`, and should stop the double render,
and hence the flash of typed content, as opposed to what is returned
from `onInputChange`.

Simplified the tests by using the existing `createComponentWithWrapper`
function, and using a stub. This should make the tests clearer.
Added a couple of tests for returning different things (e.g. null,
numbers), and one to check the new value is actually rendered in the
DOM.